pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.3",
    "opencv-python>=4.11.0",
    "pdf2image>=1.17.0",
    "Pillow>=11.2.1",
    "scipy>=1.17.1",
    "transformers==4.37.2",
    "torch",
    "tiktoken>=0.12.0",
    "torchvision>=0.21.0",
    "requests>=2.32.5",
    "verovio>=6.1.0",
    "accelerate>=1.13.0",
    "openai>=2.29.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" }
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
